+++ console/algorithms/src/pedersen/hash_uncompressed.rs
@@ -16,16 +16,16 @@
 
 use super::*;
 
-impl<G: AffineCurve, const NUM_BITS: usize> HashUncompressed for Pedersen<G, NUM_BITS> {
+impl<G: AffineCurve, const NUM_BITS: u8> HashUncompressed for Pedersen<G, NUM_BITS> {
     type Input = bool;
     type Output = G;
 
     /// Returns the Pedersen hash of the given input as an affine group element.
     fn hash_uncompressed(&self, input: &[Self::Input]) -> Result<Self::Output> {
         let mut input = Cow::Borrowed(input);
-        match input.len() <= NUM_BITS {
+        match input.len() <= NUM_BITS as usize {
             // Pad the input if it is under the required parameter size.
-            true => input.to_mut().resize(NUM_BITS, false),
+            true => input.to_mut().resize(NUM_BITS as usize, false),
             // Ensure the input size is within the parameter size,
             false => bail!("Invalid input size for Pedersen: expected <= {NUM_BITS}, found {}", input.len()),
         }
